# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Toolforge Components API
  version: 0.4.2
servers:
- url: http://127.0.0.1:8000
  description: Local direct development server.
- url: https://127.0.0.1:3000/components
  description: Lima-kilo development server.
- url: https://api.svc.tools.eqiad1.wikimedia.cloud:30003/components
  description: Toolforge internal API gateway endpoint.
- url: https://api.svc.toolsbeta.eqiad1.wikimedia.cloud:30003/components
  description: Toolforge beta internal API gateway endpoint.
- url: https://api.svc.toolforge.org/components
  description: Toolforge external API gateway endpoint.
- url: https://api.svc.beta.toolforge.org/components
  description: Toolforge beta external API gateway endpoint.
paths:
...
  /v1/tool/{toolname}/config:
    get:
      tags:
      - tool
      summary: Get Tool Config
      description: Retrieve the configuration for a specific tool.
      operationId: get_tool_config
      security:
      - APIKeyHeader: []
      parameters:
      - name: toolname
        in: path
        required: true
        schema:
          type: string
          title: Toolname
      - name: rebuild_storage
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Rebuild Storage
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_ToolConfig_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - tool
      summary: Update Tool Config
      description: Update or create the configuration for a specific tool.
      operationId: update_tool_config
      security:
      - APIKeyHeader: []
      parameters:
      - name: toolname
        in: path
        required: true
        schema:
          type: string
          title: Toolname
      - name: rebuild_storage
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Rebuild Storage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ToolConfig'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_ToolConfig_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - tool
      summary: Delete Tool Config
      description: Delete the configuration for a specific tool.
      operationId: delete_tool_config
      security:
      - APIKeyHeader: []
      parameters:
      - name: toolname
        in: path
        required: true
        schema:
          type: string
          title: Toolname
      - name: rebuild_storage
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Rebuild Storage
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_ToolConfig_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/tool/{toolname}/config/generate:
    get:
      tags:
      - tool
      summary: Generate Tool Config
      description: 'Generate the configuration for a specific tool from existing jobs
        if possible.


        Fallback to a hardcoded example otherwise. Note that this will not generate
        resources that are not supported.'
      operationId: generate_tool_config
      security:
      - APIKeyHeader: []
      parameters:
      - name: toolname
        in: path
        required: true
        schema:
          type: string
          title: Toolname
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_ToolConfig_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/tool/{toolname}/deployment/token:
    get:
      tags:
      - tool
      summary: Get Tool Deploy Token
      operationId: get_tool_deploy_token
      security:
      - APIKeyHeader: []
      parameters:
      - name: toolname
        in: path
        required: true
        schema:
          type: string
          title: Toolname
      - name: rebuild_storage
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Rebuild Storage
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_DeployToken_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - tool
      summary: Create Tool Deploy Token
      operationId: create_tool_deploy_token
      security:
      - APIKeyHeader: []
      parameters:
      - name: toolname
        in: path
        required: true
        schema:
          type: string
          title: Toolname
      - name: rebuild_storage
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Rebuild Storage
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_DeployToken_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - tool
      summary: Update Tool Deploy Token
      operationId: update_tool_deploy_token
      security:
      - APIKeyHeader: []
      parameters:
      - name: toolname
        in: path
        required: true
        schema:
          type: string
          title: Toolname
      - name: rebuild_storage
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Rebuild Storage
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_DeployToken_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - tool
      summary: Delete Tool Deploy Token
      operationId: delete_tool_deploy_token
      security:
      - APIKeyHeader: []
      parameters:
      - name: toolname
        in: path
        required: true
        schema:
          type: string
          title: Toolname
      - name: rebuild_storage
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Rebuild Storage
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_DeployToken_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/tool/{toolname}/deployment/latest:
    get:
      tags:
      - tool
      summary: Get Latest Deployment
      description: Print the latest deployment for a specific tool, sorted by creation_time
      operationId: get_latest_deployment
      security:
      - APIKeyHeader: []
      parameters:
      - name: toolname
        in: path
        required: true
        schema:
          type: string
          title: Toolname
      - name: rebuild_storage
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Rebuild Storage
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Deployment_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/tool/{toolname}/deployment/{deployment_id}/cancel:
    put:
      tags:
      - tool
      summary: Cancel Tool Deployment
      operationId: cancel_tool_deployment
      security:
      - APIKeyHeader: []
      parameters:
      - name: toolname
        in: path
        required: true
        schema:
          type: string
          title: Toolname
      - name: deployment_id
        in: path
        required: true
        schema:
          type: string
          title: Deployment Id
      - name: rebuild_storage
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Rebuild Storage
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Deployment_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/tool/{toolname}/deployment/latest/cancel:
    put:
      tags:
      - tool
      summary: Cancel Tool Deployment
      operationId: cancel_tool_deployment
      security:
      - APIKeyHeader: []
      parameters:
      - name: toolname
        in: path
        required: true
        schema:
          type: string
          title: Toolname
      - name: deployment_id
        in: query
        required: true
        schema:
          type: string
          title: Deployment Id
      - name: rebuild_storage
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Rebuild Storage
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Deployment_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/tool/{toolname}/deployment:
    get:
      tags:
      - tool
      summary: List Tool Deployments
      description: List all deployments for a specific tool.
      operationId: list_tool_deployments
      security:
      - APIKeyHeader: []
      parameters:
      - name: toolname
        in: path
        required: true
        schema:
          type: string
          title: Toolname
      - name: rebuild_storage
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Rebuild Storage
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_DeploymentList_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - tool
      summary: Create Tool Deployment
      description: Create a new tool deployment.
      operationId: create_tool_deployment
      security:
      - APIKeyHeader: []
      - APIKeyQuery: []
      parameters:
      - name: toolname
        in: path
        required: true
        schema:
          type: string
          title: Toolname
      - name: force-build
        in: query
        required: false
        schema:
          type: boolean
          title: Force Build
          description: By default the component is only rebuilt if there was a change
            in the source code and there's no build for that same commit. To force
            a re-build regardless of the source code status, use force-build
          default: false
        description: By default the component is only rebuilt if there was a change
          in the source code and there's no build for that same commit. To force a
          re-build regardless of the source code status, use force-build
      - name: force-run
        in: query
        required: false
        schema:
          type: boolean
          title: Force Run
          description: By default an already running component is not restarted if
            there are no changes to the run parameters. To force a re-run regardless
            of the run parameters, use force-run
          default: false
        description: By default an already running component is not restarted if there
          are no changes to the run parameters. To force a re-run regardless of the
          run parameters, use force-run
      - name: rebuild_storage
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Rebuild Storage
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Deployment_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/tool/{toolname}/deployment/{deployment_id}:
    delete:
      tags:
      - tool
      summary: Delete Tool Deployment
      operationId: delete_tool_deployment
      security:
      - APIKeyHeader: []
      parameters:
      - name: toolname
        in: path
        required: true
        schema:
          type: string
          title: Toolname
      - name: deployment_id
        in: path
        required: true
        schema:
          type: string
          title: Deployment Id
      - name: rebuild_storage
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Rebuild Storage
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Deployment_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - tool
      summary: Get Tool Deployment
      operationId: get_tool_deployment
      security:
      - APIKeyHeader: []
      - APIKeyQuery: []
      parameters:
      - name: toolname
        in: path
        required: true
        schema:
          type: string
          title: Toolname
      - name: deployment_id
        in: path
        required: true
        schema:
          type: string
          title: Deployment Id
      - name: rebuild_storage
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Rebuild Storage
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Deployment_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    ApiResponse_DeployToken_:
      properties:
        data:
          $ref: '#/components/schemas/DeployToken'
        messages:
          $ref: '#/components/schemas/ResponseMessages'
          default:
            info: []
            warning:
            - You are using a beta feature of Toolforge.
            error: []
      type: object
      required:
      - data
      title: ApiResponse[DeployToken]
    ApiResponse_DeploymentList_:
      properties:
        data:
          $ref: '#/components/schemas/DeploymentList'
        messages:
          $ref: '#/components/schemas/ResponseMessages'
          default:
            info: []
            warning:
            - You are using a beta feature of Toolforge.
            error: []
      type: object
      required:
      - data
      title: ApiResponse[DeploymentList]
    ApiResponse_Deployment_:
      properties:
        data:
          $ref: '#/components/schemas/Deployment'
        messages:
          $ref: '#/components/schemas/ResponseMessages'
          default:
            info: []
            warning:
            - You are using a beta feature of Toolforge.
            error: []
      type: object
      required:
      - data
      title: ApiResponse[Deployment]
...
    ApiResponse_ToolConfig_:
      properties:
        data:
          $ref: '#/components/schemas/ToolConfig'
        messages:
          $ref: '#/components/schemas/ResponseMessages'
          default:
            info: []
            warning:
            - You are using a beta feature of Toolforge.
            error: []
      type: object
      required:
      - data
      title: ApiResponse[ToolConfig]
    ContinuousComponentInfo:
      properties:
        build:
          anyOf:
          - $ref: '#/components/schemas/SourceBuildInfo'
          - $ref: '#/components/schemas/SourceBuildReference'
          title: Build
        run:
          $ref: '#/components/schemas/ContinuousRunInfo'
        component_type:
          type: string
          enum:
          - continuous
          - scheduled
          title: Component Type
          default: continuous
      type: object
      required:
      - build
      - run
      title: ContinuousComponentInfo
...
    DeployToken:
      properties:
        token:
          type: string
          format: uuid
          title: Token
        creation_date:
          type: string
          format: date-time
          title: Creation Date
      type: object
      title: DeployToken
...
    DeploymentList:
      properties:
        deployments:
          items:
            $ref: '#/components/schemas/Deployment'
          type: array
          title: Deployments
      type: object
      required:
      - deployments
      title: DeploymentList
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ResponseMessages:
      properties:
        info:
          items:
            type: string
          type: array
          title: Info
          default: []
        warning:
          items:
            type: string
          type: array
          title: Warning
          default:
          - You are using a beta feature of Toolforge.
        error:
          items:
            type: string
          type: array
          title: Error
          default: []
      type: object
      title: ResponseMessages
    ScheduledComponentInfo:
      properties:
        build:
          anyOf:
          - $ref: '#/components/schemas/SourceBuildInfo'
          - $ref: '#/components/schemas/SourceBuildReference'
          title: Build
        run:
          $ref: '#/components/schemas/ScheduledRunInfo'
        component_type:
          type: string
          enum:
          - continuous
          - scheduled
          title: Component Type
          default: scheduled
      type: object
      required:
      - build
      - run
      title: ScheduledComponentInfo
...
    SourceBuildReference:
      properties:
        reuse_from:
          type: string
          title: Reuse From
          description: Component to re-use the build from.
          examples:
          - parent-component
      type: object
      required:
      - reuse_from
      title: SourceBuildReference
    ToolConfig:
      properties:
        config_version:
          anyOf:
          - type: string
            const: v1beta1
          - type: 'null'
          title: Config Version
          default: v1beta1
          examples:
          - v1beta1
          nullable: true
        source_url:
          type: string
          minLength: 1
          format: uri
          title: Source Url
          description: If passed, it will ignore anything else and try to fetch the
            configuration from the given URL on every deploy.
        components:
          additionalProperties:
            anyOf:
            - $ref: '#/components/schemas/ContinuousComponentInfo'
            - $ref: '#/components/schemas/ScheduledComponentInfo'
          type: object
          minProperties: 1
          title: Components
          description: List of components to run. Each component matches a continuous
            job, scheduled job, one-off job or webservice.
      type: object
      required:
      - components
      title: ToolConfig
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    APIKeyHeader:
      type: apiKey
      in: header
      name: x-toolforge-tool
    APIKeyQuery:
      type: apiKey
      in: query
      name: token